+++ raesan-common/src/schema.rs
@@ -8,8 +8,8 @@ diesel::table! {
         subject_id -> Text,
         subject_name -> Text,
         class_name -> Integer,
-        created_at -> Integer,
-        updated_at -> Integer,
+        created_at -> BigInt,
+        updated_at -> BigInt,
     }
 }
 
@@ -17,8 +17,8 @@ diesel::table! {
     classes (id) {
         id -> Text,
         name -> Integer,
-        created_at -> Integer,
-        updated_at -> Integer,
+        created_at -> BigInt,
+        updated_at -> BigInt,
     }
 }
 
@@ -30,8 +30,8 @@ diesel::table! {
         subject_name -> Text,
         class_name -> Integer,
         chapter_id -> Text,
-        created_at -> Integer,
-        updated_at -> Integer,
+        created_at -> BigInt,
+        updated_at -> BigInt,
     }
 }
 
@@ -42,8 +42,8 @@ diesel::table! {
         display_name -> Text,
         class_id -> Text,
         class_name -> Integer,
-        created_at -> Integer,
-        updated_at -> Integer,
+        created_at -> BigInt,
+        updated_at -> BigInt,
     }
 }
