+++ raesan/src/core/models.rs
@@ -19,8 +19,8 @@ use serde;
 pub struct Class {
     pub id: String,
     pub name: i32,
-    pub created_at: i32,
-    pub updated_at: i32,
+    pub created_at: i64,
+    pub updated_at: i64,
 }
 
 // ----- `Subject` model struct
@@ -42,8 +42,8 @@ pub struct Subject {
     pub display_name: String,
     pub class_id: String,
     pub class_name: i32,
-    pub created_at: i32,
-    pub updated_at: i32,
+    pub created_at: i64,
+    pub updated_at: i64,
 }
 
 // ----- `Chapter` model struct
@@ -66,8 +66,8 @@ pub struct Chapter {
     pub subject_id: String,
     pub subject_name: String,
     pub class_name: i32,
-    pub created_at: i32,
-    pub updated_at: i32,
+    pub created_at: i64,
+    pub updated_at: i64,
 }
 
 // ----- `Question` model struct
@@ -90,8 +90,8 @@ pub struct Question {
     pub subject_name: String,
     pub class_name: i32,
     pub chapter_id: String,
-    pub created_at: i32,
-    pub updated_at: i32,
+    pub created_at: i64,
+    pub updated_at: i64,
 }
 
 // ----- `CreateTestInput` struct
